pull up {Default,}Fetcher to its own package
