+++ schema_salad/ref_resolver.py
@@ -2,20 +2,17 @@ import copy
 import logging
 import os
 import re
-import sys
 import xml.sax
-from io import StringIO, open
+from io import StringIO
 from typing import (
     Any,
     Callable,
     Dict,
-    Iterable,
     List,
     MutableMapping,
     MutableSequence,
     Optional,
     Set,
-    Tuple,
     TypeVar,
     Union,
     cast,
@@ -33,25 +30,23 @@ from ruamel import yaml
 from ruamel.yaml.comments import CommentedMap, CommentedSeq, LineCol
 
 from .exceptions import SchemaSaladException, ValidationException
+from .fetcher import DefaultFetcher
 from .sourceline import SourceLine, add_lc_filename, relname
-from .utils import aslist, onWindows
+from .utils import (
+    AttachmentsType,
+    CacheType,
+    ContextType,
+    FetcherCallableType,
+    IdxResultType,
+    IdxType,
+    ResolvedRefType,
+    ResolveType,
+    aslist,
+    onWindows,
+)
 
 _logger = logging.getLogger("salad")
-ContextType = Dict[str, Union[Dict[str, Any], str, Iterable[str]]]
-DocumentType = TypeVar("DocumentType", CommentedSeq, CommentedMap)
-DocumentOrStrType = TypeVar("DocumentOrStrType", CommentedSeq, CommentedMap, str)
-FieldType = TypeVar("FieldType", str, CommentedSeq, CommentedMap)
-ResolveType = Union[int, float, str, CommentedMap, CommentedSeq, None]
-ResolvedRefType = Tuple[ResolveType, CommentedMap]
-IdxResultType = Union[CommentedMap, CommentedSeq, str, None]
-IdxType = Dict[str, IdxResultType]
-CacheType = Dict[str, Union[str, Graph, bool]]
-FetcherCallableType = Callable[[CacheType, requests.sessions.Session], "Fetcher"]
-AttachementsSig = Callable[[Union[CommentedMap, CommentedSeq]], bool]
 typeDSLregex = re.compile(str(r"^([^[?]+)(\[\])?(\?)?$"))
-
-
-_re_drive = re.compile(r"/([a-zA-Z]):")
 
 
 def file_uri(path: str, split_frag: bool = False) -> str:
@@ -100,8 +95,7 @@ def to_validation_exception(e: yaml.error.MarkedYAMLError) -> ValidationExceptio
         parent.end = None
         parent.children = [exc]
         return parent
-    else:
-        return exc
+    return exc
 
 
 class NormDict(Dict[str, Union[CommentedMap, CommentedSeq, str, None]]):
@@ -142,184 +136,6 @@ def SubLoader(loader: "Loader") -> "Loader":
     )
 
 
-class Fetcher(object):
-    def __init__(
-        self, cache: CacheType, session: Optional[requests.sessions.Session],
-    ) -> None:
-        pass
-
-    def fetch_text(self, url: str) -> str:
-        raise NotImplementedError()
-
-    def check_exists(self, url: str) -> bool:
-        raise NotImplementedError()
-
-    def urljoin(self, base_url: str, url: str) -> str:
-        raise NotImplementedError()
-
-    schemes = ["file", "http", "https", "mailto"]
-
-    def supported_schemes(self) -> List[str]:
-        return self.schemes
-
-
-class DefaultFetcher(Fetcher):
-    def __init__(
-        self, cache: CacheType, session: Optional[requests.sessions.Session],
-    ) -> None:
-        self.cache = cache
-        self.session = session
-
-    def fetch_text(self, url: str) -> str:
-        if url in self.cache and self.cache[url] is not True:
-            # treat "True" as a placeholder that indicates something exists but
-            # not necessarily what its contents is.
-            result = self.cache[url]
-            assert isinstance(result, str)
-            return result
-
-        split = urllib.parse.urlsplit(url)
-        scheme, path = split.scheme, split.path
-
-        if scheme in ["http", "https"] and self.session is not None:
-            try:
-                resp = self.session.get(url)
-                resp.raise_for_status()
-            except Exception as e:
-                raise ValidationException("Error fetching {}: {}".format(url, e)) from e
-            return resp.text
-        if scheme == "file":
-            try:
-                # On Windows, url.path will be /drive:/path ; on Unix systems,
-                # /path. As we want drive:/path instead of /drive:/path on Windows,
-                # remove the leading /.
-                if os.path.isabs(
-                    path[1:]
-                ):  # checking if pathis valid after removing front / or not
-                    path = path[1:]
-                with open(
-                    urllib.request.url2pathname(str(path)), encoding="utf-8"
-                ) as fp:
-                    return str(fp.read())
-
-            except OSError as err:
-                if err.filename == path:
-                    raise ValidationException(str(err)) from err
-                else:
-                    raise ValidationException(
-                        "Error reading {}: {}".format(url, err)
-                    ) from err
-        raise ValidationException("Unsupported scheme in url: {}".format(url))
-
-    def check_exists(self, url: str) -> bool:
-        if url in self.cache:
-            return True
-
-        split = urllib.parse.urlsplit(url)
-        scheme, path = split.scheme, split.path
-
-        if scheme in ["http", "https"] and self.session is not None:
-            try:
-                resp = self.session.head(url)
-                resp.raise_for_status()
-            except Exception:
-                return False
-            self.cache[url] = True
-            return True
-        if scheme == "file":
-            return os.path.exists(urllib.request.url2pathname(str(path)))
-        if scheme == "mailto":
-            return True
-        raise ValidationException("Unsupported scheme in url: {}".format(url))
-
-    def urljoin(self, base_url: str, url: str) -> str:
-        if url.startswith("_:"):
-            return url
-
-        basesplit = urllib.parse.urlsplit(base_url)
-        split = urllib.parse.urlsplit(url)
-        if basesplit.scheme and basesplit.scheme != "file" and split.scheme == "file":
-            raise ValidationException(
-                "Not resolving potential remote exploit {} from base {}".format(
-                    url, base_url
-                )
-            )
-
-        if sys.platform == "win32":
-            if base_url == url:
-                return url
-            basesplit = urllib.parse.urlsplit(base_url)
-            # note that below might split
-            # "C:" with "C" as URI scheme
-            split = urllib.parse.urlsplit(url)
-
-            has_drive = split.scheme and len(split.scheme) == 1
-
-            if basesplit.scheme == "file":
-                # Special handling of relative file references on Windows
-                # as urllib seems to not be quite up to the job
-
-                # netloc MIGHT appear in equivalents of UNC Strings
-                # \\server1.example.com\path as
-                # file:///server1.example.com/path
-                # https://tools.ietf.org/html/rfc8089#appendix-E.3.2
-                # (TODO: test this)
-                netloc = split.netloc or basesplit.netloc
-
-                # Check if url is a local path like "C:/Users/fred"
-                # or actually an absolute URI like http://example.com/fred
-                if has_drive:
-                    # Assume split.scheme is actually a drive, e.g. "C:"
-                    # so we'll recombine into a path
-                    path_with_drive = urllib.parse.urlunsplit(
-                        (split.scheme, "", split.path, "", "")
-                    )
-                    # Compose new file:/// URI with path_with_drive
-                    # .. carrying over any #fragment (?query just in case..)
-                    return urllib.parse.urlunsplit(
-                        ("file", netloc, path_with_drive, split.query, split.fragment)
-                    )
-                if (
-                    not split.scheme
-                    and not netloc
-                    and split.path
-                    and split.path.startswith("/")
-                ):
-                    # Relative - but does it have a drive?
-                    base_drive = _re_drive.match(basesplit.path)
-                    drive = _re_drive.match(split.path)
-                    if base_drive and not drive:
-                        # Keep drive letter from base_url
-                        # https://tools.ietf.org/html/rfc8089#appendix-E.2.1
-                        # e.g. urljoin("file:///D:/bar/a.txt", "/foo/b.txt")
-                        #          == file:///D:/foo/b.txt
-                        path_with_drive = "/{}:{}".format(
-                            base_drive.group(1), split.path
-                        )
-                        return urllib.parse.urlunsplit(
-                            (
-                                "file",
-                                netloc,
-                                path_with_drive,
-                                split.query,
-                                split.fragment,
-                            )
-                        )
-
-                # else: fall-through to resolve as relative URI
-            elif has_drive:
-                # Base is http://something but url is C:/something - which urllib
-                # would wrongly resolve as an absolute path that could later be used
-                # to access local files
-                raise ValidationException(
-                    "Not resolving potential remote exploit {} from base {}".format(
-                        url, base_url
-                    )
-                )
-
-        return urllib.parse.urljoin(base_url, url)
-
-
 class Loader(object):
     def __init__(
         self,
@@ -332,7 +148,7 @@ class Loader(object):
         fetcher_constructor: Optional[FetcherCallableType] = None,
         skip_schemas: Optional[bool] = None,
         url_fields: Optional[Set[str]] = None,
-        allow_attachments: Optional[AttachementsSig] = None,
+        allow_attachments: Optional[AttachmentsType] = None,
         doc_cache: Union[str, bool] = True,
     ) -> None:
 
@@ -461,14 +277,13 @@ class Loader(object):
 
         if vocab_term and url in self.rvocab:
             return self.rvocab[url]
-        else:
-            return url
+        return url
 
     def _add_properties(self, s: str) -> None:
         for _, _, rng in self.graph.triples((s, RDFS.range, None)):
             literal = (
                 str(rng).startswith("http://www.w3.org/2001/XMLSchema#")
-                and not str(rng) == "http://www.w3.org/2001/XMLSchema#anyURI"
+                and str(rng) != "http://www.w3.org/2001/XMLSchema#anyURI"
             ) or str(rng) == "http://www.w3.org/2000/01/rdf-schema#Literal"
             if not literal:
                 self.url_fields.add(str(s))
