+++ schema_salad/tests/test_fetch.py
@@ -1,17 +1,80 @@
 import os
+from typing import Optional
+from urllib.parse import urljoin, urlsplit
 
 import pytest  # type: ignore
+import requests
 
-import schema_salad.main
-import schema_salad.ref_resolver
-import schema_salad.schema
+from schema_salad.fetcher import Fetcher
+from schema_salad.ref_resolver import Loader, file_uri
+from schema_salad.utils import CacheType
 
-from .other_fetchers import testFetcher
+
+class testFetcher(Fetcher):
+    def __init__(
+        self, cache: CacheType, session: Optional[requests.sessions.Session],
+    ) -> None:
+        pass
+
+    def fetch_text(self, url: str) -> str:
+        if url == "keep:abc+123/foo.txt":
+            return "hello: keepfoo"
+        if url.endswith("foo.txt"):
+            return "hello: foo"
+        else:
+            raise RuntimeError("Not foo.txt")
+
+    def check_exists(self, url: str) -> bool:
+        if url.endswith("foo.txt"):
+            return True
+        else:
+            return False
+
+    def urljoin(self, base: str, url: str) -> str:
+        urlsp = urlsplit(url)
+        if urlsp.scheme:
+            return url
+        basesp = urlsplit(base)
+
+        if basesp.scheme == "keep":
+            return base + "/" + url
+        return urljoin(base, url)
+
+
+class CWLTestFetcher(Fetcher):
+    def __init__(
+        self, cache: CacheType, session: Optional[requests.sessions.Session],
+    ) -> None:
+        pass
+
+    def fetch_text(self, url):  # type: (str) -> str
+        if url == "baz:bar/foo.cwl":
+            return """
+cwlVersion: v1.0
+class: CommandLineTool
+baseCommand: echo
+inputs: []
+outputs: []
+"""
+        raise RuntimeError("Not foo.cwl, was %s" % url)
+
+    def check_exists(self, url):  # type: (str) -> bool
+        return url == "baz:bar/foo.cwl"
+
+    def urljoin(self, base: str, url: str) -> str:
+        urlsp = urlsplit(url)
+        if urlsp.scheme:
+            return url
+        basesp = urlsplit(base)
+
+        if basesp.scheme == "keep":
+            return base + "/" + url
+        return urljoin(base, url)
 
 
 def test_fetcher() -> None:
 
-    loader = schema_salad.ref_resolver.Loader({}, fetcher_constructor=testFetcher)
+    loader = Loader({}, fetcher_constructor=testFetcher)
     assert {"hello": "foo"} == loader.resolve_ref("foo.txt")[0]
     assert {"hello": "keepfoo"} == loader.resolve_ref(
         "foo.txt", base_url="keep:abc+123"
@@ -24,9 +87,9 @@ def test_fetcher() -> None:
 
 
 def test_cache() -> None:
-    loader = schema_salad.ref_resolver.Loader({})
+    loader = Loader({})
     foo = os.path.join(os.getcwd(), "foo.txt")
-    foo = schema_salad.ref_resolver.file_uri(foo)
+    foo = file_uri(foo)
     loader.cache.update({foo: "hello: foo"})
     print(loader.cache)
     assert {"hello": "foo"} == loader.resolve_ref("foo.txt")[0]
